Move action_log related interface methods into BMInterfaceGameAction in preparation for future work
